requirements.txt
flask==3.1.1
flask-cors==5.0.1
flask-sqlalchemy==3.1.1
google-generativeai==0.8.0
pdfplumber==0.11.0
python-dotenv==1.1.0
gunicorn==23.0.0
